pyproject.toml
[tool.poetry]
name = "Video3D"
version = "0.1.0"
description = ""
authors = ["ytya <ytyadev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
torch = { url="https://download.pytorch.org/whl/cu117/torch-2.0.1%2Bcu117-cp310-cp310-win_amd64.whl" }
torchvision = { url="https://download.pytorch.org/whl/cu117/torchvision-0.15.2%2Bcu117-cp310-cp310-win_amd64.whl" }
tqdm = "*"
pillow = "*"
fire = "*"
opencv-python = "*"
scikit-video = "*"
einops = "*"
matplotlib = "*"

# MiDaS
timm = "0.6.12" # 0.9.1 だと実行時エラー
imutils = "0.5.4"

# frame-interpolation
# tensorflow = "2.11.1"
# tensorflow-io-gcs-filesystem = "0.31.0"  # 0.32.0 だとインストールエラー
# opencv-contrib-python = "*"
# matplotlib = "*"
# numpy = "*"
# einops = "*"
# imutils = "*"
# tensorflow = "2.8.4"
# tensorflow-datasets = "*"
# tensorflow-addons = "*"
# absl-py = "*"
# gin-config = "*"
# parameterized = "*"
# mediapy = "*"
# scikit-image = "*"
# apache-beam = "2.44.0"
# natsort = "*"
# gdown = "*"
# pyarrow = "9.0.0"


[tool.poetry.dev-dependencies]
jupyterlab = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py310"
line-length = 119
extend-exclude = [
  "NVDS",
]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "I",
]
ignore = [
  "E501", # Line too long
  "W293", # Blank line contains whitespace
]
fixable = ["ALL"]
